// assets/js/interfaces.ts
/**
 * Copyright (C) 2018 Johnny Robeson <johnny@localmomentum.net>
 *
 * This program is free software: you can redistribute it and/or modify
 * it under the terms of the GNU Affero General Public License as
 * published by the Free Software Foundation, either version 3 of the
 * License, or (at your option) any later version.
 *
 * This program is distributed in the hope that it will be useful,
 * but WITHOUT ANY WARRANTY; without even the implied warranty of
 * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 * GNU Affero General Public License for more details.
 *
 * You should have received a copy of the GNU Affero General Public License
 * along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */

export interface UnitInterface {
  id: number;
  zones: number;
}

export interface PlayerInterface {
  id: number;
  name: string;
  hitPoints: number;
  score: number;
  zoneHits: number[];
  radioId: string;
  unit: UnitInterface;
}

export interface TeamInterface {
  id: number;
  name: string;
  color: string;
  players: PlayerInterface[];
  score?: number;
  hitPoints?: number;
}

export interface GameInterface {
  gameType: 'target' | 'team';
  teams: TeamInterface[];
  active: boolean;
  endsAt: string; // iso 6601
  timeTotal: number;
}